update1

+++ file.go
@@ -19,9 +19,9 @@ import (
 	"sync"
 	"time"
 
-	"github.com/camlistore/lock"
-	"github.com/cznic/exp/lldb"
+	"github.com/cznic/lldb"
 	"github.com/cznic/mathutil"
+	"github.com/cznic/ql/vendored/github.com/camlistore/go4/lock"
 )
 
 const (
@@ -89,7 +89,7 @@ func OpenFile(name string, opt *Options) (db *DB, err error) {
 		}
 	}
 
-	fi, err := newFileFromOSFile(f) // always ACID
+	fi, err := newFileFromOSFile(f, opt.Headroom) // always ACID
 	if err != nil {
 		return
 	}
@@ -126,10 +126,19 @@ func OpenFile(name string, opt *Options) (db *DB, err error) {
 // interface.
 //
 // If TempFile is nil it defaults to ioutil.TempFile.
+//
+// Headroom
+//
+// Headroom selecst the minimum size a WAL file will have. The "extra"
+// allocated file space serves as a headroom. Commits that fit into the
+// headroom should not fail due to 'not enough space on the volume' errors. The
+// headroom parameter is first rounded-up to a non negative multiple of the
+// size of the lldb.Allocator atom.
 type Options struct {
 	CanCreate bool
 	OSFile    lldb.OSFile
 	TempFile  func(dir, prefix string) (f lldb.OSFile, err error)
+	Headroom  int64
 }
 
 type fileBTreeIterator struct {
@@ -386,7 +395,7 @@ type file struct {
 	wal      *os.File
 }
 
-func newFileFromOSFile(f lldb.OSFile) (fi *file, err error) {
+func newFileFromOSFile(f lldb.OSFile, headroom int64) (fi *file, err error) {
 	nm := lockName(f.Name())
 	lck, err := lock.Lock(nm)
 	if err != nil {
@@ -409,7 +418,7 @@ func newFileFromOSFile(f lldb.OSFile) (fi *file, err error) {
 	w, err = os.OpenFile(wn, os.O_CREATE|os.O_EXCL|os.O_RDWR, 0666)
 	closew = true
 	defer func() {
-		if closew {
+		if w != nil && closew {
 			nm := w.Name()
 			w.Close()
 			os.Remove(nm)
@@ -434,9 +443,7 @@ func newFileFromOSFile(f lldb.OSFile) (fi *file, err error) {
 			return nil, err
 		}
 
-		if st.Size() != 0 {
-			return nil, fmt.Errorf("(file-001) non empty WAL file %s exists", wn)
-		}
+		closew = st.Size() == 0
 	}
 
 	info, err := f.Stat()
@@ -454,7 +461,7 @@ func newFileFromOSFile(f lldb.OSFile) (fi *file, err error) {
 
 		filer := lldb.Filer(lldb.NewOSFiler(f))
 		filer = lldb.NewInnerFiler(filer, 16)
-		if filer, err = lldb.NewACIDFiler(filer, w); err != nil {
+		if filer, err = lldb.NewACIDFiler(filer, w, lldb.MinWAL(headroom)); err != nil {
 			return nil, err
 		}
 
@@ -508,7 +515,7 @@ func newFileFromOSFile(f lldb.OSFile) (fi *file, err error) {
 
 		filer := lldb.Filer(lldb.NewOSFiler(f))
 		filer = lldb.NewInnerFiler(filer, 16)
-		if filer, err = lldb.NewACIDFiler(filer, w); err != nil {
+		if filer, err = lldb.NewACIDFiler(filer, w, lldb.MinWAL(headroom)); err != nil {
 			return nil, err
 		}
 
@@ -554,7 +561,7 @@ func (s *file) OpenIndex(unique bool, handle int64) (btreeIndex, error) {
 		return nil, err
 	}
 
-	return &fileIndex{s, handle, t, unique}, nil
+	return &fileIndex{s, handle, t, unique, newGobCoder()}, nil
 }
 
 func (s *file) CreateIndex(unique bool) ( /* handle */ int64, btreeIndex, error) {
@@ -563,7 +570,7 @@ func (s *file) CreateIndex(unique bool) ( /* handle */ int64, btreeIndex, error)
 		return -1, nil, err
 	}
 
-	return h, &fileIndex{s, h, t, unique}, nil
+	return h, &fileIndex{s, h, t, unique, newGobCoder()}, nil
 }
 
 func (s *file) Acid() bool { return s.wal != nil }
@@ -795,7 +802,7 @@ func (s *file) Read(dst []interface{}, h int64, cols ...*col) (data []interface{
 
 	for _, col := range cols {
 		i := col.index + 2
-		if i >= len(rec) {
+		if i >= len(rec) || rec[i] == nil {
 			continue
 		}
 
@@ -825,8 +832,6 @@ func (s *file) Read(dst []interface{}, h int64, cols ...*col) (data []interface{
 		case qUint64:
 		case qBlob, qBigInt, qBigRat, qTime, qDuration:
 			switch x := rec[i].(type) {
-			case nil:
-				rec[i] = nil
 			case []byte:
 				rec[i] = chunk{f: s, b: x}
 			default:
@@ -1098,6 +1103,7 @@ type fileIndex struct {
 	h      int64
 	t      *lldb.BTree
 	unique bool
+	codec  *gobCoder
 }
 
 func (x *fileIndex) Clear() error {
@@ -1125,6 +1131,13 @@ func isIndexNull(data []interface{}) bool {
 // The []byte version of the key in the BTree shares chunks, if any, with
 // the value stored in the record.
 func (x *fileIndex) Create(indexedValues []interface{}, h int64) error {
+	for i, indexedValue := range indexedValues {
+		chunk, ok := indexedValue.(chunk)
+		if ok {
+			indexedValues[i] = chunk.b
+		}
+	}
+
 	t := x.t
 	switch {
 	case !x.unique:
@@ -1197,8 +1210,51 @@ func (x *fileIndex) Drop() error {
 	return x.f.a.Free(x.h)
 }
 
-func (x *fileIndex) Seek(indexedValues []interface{}) (indexIterator, bool, error) { //TODO(indices) blobs: +test
-	k, err := lldb.EncodeScalars(append(indexedValues, 0)...)
+// []interface{}{qltype, ...}->[]interface{}{lldb scalar type, ...}
+func (x *fileIndex) flatten(data []interface{}) (err error) {
+	for i, v := range data {
+		tag := 0
+		var b []byte
+		switch xx := v.(type) {
+		case []byte:
+			tag = qBlob
+			b = xx
+		case *big.Int:
+			tag = qBigInt
+			b, err = x.codec.encode(xx)
+		case *big.Rat:
+			tag = qBigRat
+			b, err = x.codec.encode(xx)
+		case time.Time:
+			tag = qTime
+			b, err = x.codec.encode(xx)
+		case time.Duration:
+			tag = qDuration
+			b, err = x.codec.encode(xx)
+		default:
+			continue
+		}
+		if err != nil {
+			return
+		}
+
+		var buf []byte
+		if buf, err = lldb.EncodeScalars([]interface{}{tag, b}...); err != nil {
+			return
+		}
+
+		data[i] = buf
+	}
+	return
+}
+
+func (x *fileIndex) Seek(indexedValues []interface{}) (indexIterator, bool, error) {
+	data := append(indexedValues, 0)
+	if err := x.flatten(data); err != nil {
+		return nil, false, err
+	}
+
+	k, err := lldb.EncodeScalars(data...)
 	if err != nil {
 		return nil, false, err
 	}
